Suggestions n' Stuff
Added some better images for the homepage, and added Enterprise’s logo
at the bottom of the page. On the sponsor page, Enterprise’s address
was included (as per their request), and fixed the repeat name on the
list of razoo patrons. For both the home and sponsor page, I removed
the calendar option from the navbar.