Making RobotsPage Mobile-friendly
Added “cards” to RobotsPage (each of the entries acts as a responsive
div), and made the video remain centered regardless of screen size.
Also, the content was offset by a bit more to better show the
background video. P.S. Took way longer than I thought it would to fix
the video.