Bye Bye bad javascript
Got rid of the terrible javascript that was used to counteract the
expansion of the Navbar in compact view (mobile), and put a media query
in its place. Also, the outreach page video sticks to the left side
now, although more changes will be made to make the video and the rest
of the page more responsive.