Outreach Fixes
Added a meta tag to the outreach page, so the page now works correctly
on mobile devices. Also, did edits on the content of the page, although
more is likely to be changed. TWIMC, the playsinline tag was
intentionally left out of the video on this page (it just doesn’t look
great because the video is pretty short and the kid that catches the
ball is not visible).