Added Lemonaide Day
Added the Lemonaide day entry, and moved clearfix to the css file for
the menubar (index_new.css) so it will be accessible anywhere. Also,
moved all outreach images to a dedicated folder and fixed the Barnes
and Noble Mini Maker Faire so it’s not so gosh-darn ugly. Curse you
Brendan of the past.